+++ smartsheet/models/enums/approver_type.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+class ApproverType(str, Enum):
+    GROUPS = 'GROUPS'
+    USERS = 'USERS'
+    WORKSPACE_ADMINS = 'WORKSPACE_ADMINS'