+++ smartsheet/models/enums/downgrade_approval_mode.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+class DowngradeApprovalMode(str, Enum):
+    NONE = 'NONE'
+    APPROVAL_NEEDED = 'APPROVAL_NEEDED'
+    CUSTOM = 'CUSTOM'